+++ .gitignore
@@ -18,3 +18,6 @@ toolchain.cmake
 
 # Users often install under the root directory
 install/
+
+# Python files
+__pycache__/
